requirements.txt
djangorestframework==3.17.1
asgiref==3.11.1
Django==6.0.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
